requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
colpali-engine>=0.2.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
autogen-agentchat>=0.2.0
autogen-ext>=0.2.0
autogen-core>=0.2.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
